backend/requirements.txt
Flask==2.1.0
Flask-CORS==3.0.10
Flask-SQLAlchemy==2.5.1
PyJWT==2.3.0
requests==2.27.1
python-dotenv==0.19.2
gunicorn==20.1.0
openai==1.3.5

